+++ externals/kyuubi-spark-sql-engine/src/test/scala/org/apache/kyuubi/jdbc/KyuubiHiveDriverSuite.scala
@@ -24,7 +24,6 @@ import org.apache.kyuubi.IcebergSuiteMixin
 import org.apache.kyuubi.engine.spark.WithSparkSQLEngine
 import org.apache.kyuubi.engine.spark.shim.SparkCatalogShim
 import org.apache.kyuubi.jdbc.hive.{KyuubiConnection, KyuubiStatement}
-import org.apache.kyuubi.jdbc.hive.KyuubiSQLException
 import org.apache.kyuubi.tags.IcebergTest
 
 @IcebergTest
@@ -116,30 +115,6 @@ class KyuubiHiveDriverSuite extends WithSparkSQLEngine with IcebergSuiteMixin {
         statement.executeScala(code)
       }.getMessage
       assert(e.contains("Query timed out"))
-    } finally {
-      statement.close()
-      connection.close()
-    }
-  }
-
-  test("executePython support timeout") {
-    val driver = new KyuubiHiveDriver()
-    val connection = driver.connect(getJdbcUrl, new Properties())
-    val statement = connection.createStatement().asInstanceOf[KyuubiStatement]
-    statement.setQueryTimeout(5)
-    try {
-      var code =
-        """
-          |import time
-          |time.sleep(10)
-          |""".stripMargin
-      var e = intercept[SQLTimeoutException] {
-        statement.executePython(code)
-      }.getMessage
-      assert(e.contains("Query timed out"))
-      code = "bad_code"
-      e = intercept[KyuubiSQLException](statement.executePython(code)).getMessage
-      assert(e.contains("Interpret error"))
     } finally {
       statement.close()
       connection.close()
